pw/requirements.txt
discord.py>=2.3.0
asyncpg>=0.28.0
cryptography>=41.0.0
python-dotenv>=1.0.0